buildAMI: bake an AMI without using ciinabox-bakery.
This moves  the configuration and baking of AMIs with Packer from ciinabox-bakery into the pipeline. The 'configure' and 'bakery' from the repository scripts are now combined and simplified into this pipeline step.

Note that it uses a different repository for the Packer templates, as these changes are not backwards compatible since the Packer templates themselves have been modified.